Restrucuring and renaming in semi-explicit eqlb.
