+++ cpp/dolfinx_eqlb/se/CMakeLists.txt
@@ -3,11 +3,13 @@ set(HEADERS_se
     ${CMAKE_CURRENT_SOURCE_DIR}/Patch.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/PatchData.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/ProblemData.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/assemble_patch_semiexplt.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/assembly.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/deqlb_se.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/fluxmin_kernel.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/reconstruction.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/solve_patch_semiexplt.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/solve_patch_weaksym.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/stressmin_kernel.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/utils.hpp
     PARENT_SCOPE
 )
